Remove test_sorting.py
