+++ python/cudf/cudf/tests/series/methods/test_sort_values.py
@@ -18,3 +18,12 @@ def test_string_sort(data, ascending):
     got = gs.sort_values(ascending=ascending)
 
     assert_eq(expect, got)
+
+
+def test_series_sort_values_ignore_index(ignore_index):
+    gsr = cudf.Series([1, 3, 5, 2, 4])
+    psr = gsr.to_pandas()
+
+    expect = psr.sort_values(ignore_index=ignore_index)
+    got = gsr.sort_values(ignore_index=ignore_index)
+    assert_eq(expect, got)
